No need to distribute task when no suitable builders are available.
If a package is only for i386 machine and there is not i386 builder
it means the package will not be built, so don't bother distributing
the task and uploading the source package. However, if the package
is for i386 and am64 while there is only i386 builder available
(i.e. at least one builder is available) then proceed as usual.